switch rendering to opengl instead of sdl2 textures
Has a few caveats:
- still very messy
- libRocket doesn't work, so no GUI

+++ components/render/render.h
@@ -79,7 +79,7 @@ namespace Render
 
     void draw();
 
-    void drawAt(const Sprite& sprite, size_t x, size_t y); 
+    void drawAt(const Sprite& sprite, int32_t x, int32_t y);
 
     class SpriteGroup
     {
